renamed rawtextparser #660
