+++ analysis/import/RawTextParser/src/test/resources/cc_projects/project_1.cc.json
