+++ analysis/import/RawTextParser/src/test/resources/sampleproject/tabs.xyz
